% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.calculate.R
\name{acc.calculate}
\alias{acc.calculate}
\title{ACC calculate}
\usage{
acc.calculate(network_table, ground_truth)
}
\arguments{
\item{network_table}{The weight data table of network}

\item{ground_truth}{Ground truth for calculate AUC}
}
\value{
ACC value
}
\description{
ACC calculate
}
\examples{
data("example_matrix")
data("example_ground_truth")
network_table <- inferCSN(example_matrix)
acc.calculate(network_table, example_ground_truth)
}
