% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.calculate.R
\name{auc.calculate}
\alias{auc.calculate}
\title{auc.calculate}
\usage{
auc.calculate(
  weightList = NULL,
  groundTruth = NULL,
  groundTruthTPath = NULL,
  plot = FALSE,
  fileSave = NULL
)
}
\arguments{
\item{weightList}{weightList}

\item{groundTruth}{groundTruth}

\item{groundTruthTPath}{groundTruthTPath}

\item{plot}{Logic value}

\item{fileSave}{The figure name}
}
\value{
AUC values and AUC plot
}
\description{
AUC value calculate
}
\examples{
data("exampleDataMatrix")
data("exampleDataGroundTruth")
weightList <- inferCSN(exampleDataMatrix, cores = 1, verbose = TRUE, algorithm = "CDPSI")
auc <- auc.calculate(weightList, exampleDataGroundTruth, plot = TRUE)
head(auc)
}
