% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.calculate.R
\name{auprc.curve}
\alias{auprc.curve}
\title{auprc.curve
 auprc}
\usage{
auprc.curve(pred, obs, plot = plot)
}
\arguments{
\item{pred}{pred}

\item{obs}{obs}

\item{plot}{plot}
}
\value{
auprc
}
\description{
auprc.curve
 auprc
}
