% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.calculate.R
\name{auroc.curve}
\alias{auroc.curve}
\title{auroc.curve
 auroc}
\usage{
auroc.curve(pred, obs, plot = plot)
}
\arguments{
\item{pred}{pred}

\item{obs}{obs}

\item{plot}{plot}
}
\value{
auroc
}
\description{
auroc.curve
 auroc
}
