% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.gene.rank.R
\name{calculate.gene.rank}
\alias{calculate.gene.rank}
\title{Calculate and rank TFs in network}
\usage{
calculate.gene.rank(
  network_table,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)
}
\arguments{
\item{network_table}{The weight data table of network.}

\item{regulators}{Regulators list.}

\item{targets}{Targets list.}

\item{directed}{If network is directed or not.}
}
\value{
A data.table with three columns
}
\description{
Calculate and rank TFs in network
}
\examples{
data("example_matrix")
network_table <- inferCSN(example_matrix)
head(calculate.gene.rank(network_table))
head(calculate.gene.rank(network_table, regulators = "g1"))
}
