% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_rank.R
\name{calculate_gene_rank}
\alias{calculate_gene_rank}
\title{Rank TFs and genes in network}
\usage{
calculate_gene_rank(
  network_table,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)
}
\arguments{
\item{network_table}{Network edge table.}

\item{regulators, targets}{Nodes to include.}

\item{directed}{Whether the network is directed.}
}
\value{
A table of gene rank.
}
\description{
Rank TFs and genes in network
}
\examples{
data(example_matrix)
network_table <- inferCSN(example_matrix)
head(calculate_gene_rank(network_table))
head(calculate_gene_rank(network_table, regulators = "g1"))
head(calculate_gene_rank(network_table, targets = "g1"))
}
