% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.arguments}
\alias{check.arguments}
\title{Check input arguments}
\usage{
check.arguments(
  matrix,
  penalty,
  algorithm,
  crossValidation,
  nFolds,
  kFolds,
  rThreshold,
  regulators,
  targets,
  maxSuppSize,
  verbose,
  cores
)
}
\arguments{
\item{matrix}{An expression matrix, cells by genes}

\item{penalty}{The type of regularization.
This can take either one of the following choices: "L0"and "L0L2".
For high-dimensional and sparse data, such as single-cell transcriptome data, "L0L2" is more effective}

\item{algorithm}{Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time}

\item{crossValidation}{Check whether cross validation is used}

\item{nFolds}{The number of folds for cross-validation}

\item{kFolds}{The number of folds for sample split}

\item{rThreshold}{rThreshold}

\item{regulators}{Regulator genes}

\item{targets}{Target genes}

\item{maxSuppSize}{The number of non-zore coef, this value will affect the final performance}

\item{verbose}{Print detailed information}

\item{cores}{CPU cores}
}
\description{
Check input arguments
}
