% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.inferCSN}
\alias{coef.inferCSN}
\alias{coef.inferCSNCV}
\title{Extract Solutions}
\usage{
\method{coef}{inferCSN}(object, lambda = NULL, gamma = NULL, supportSize = NULL, ...)

\method{coef}{inferCSNCV}(object, lambda = NULL, gamma = NULL, ...)
}
\arguments{
\item{object}{The output of inferCSN.fit or inferCSN.cvfit}

\item{lambda}{The value of lambda at which to extract the solution.}

\item{gamma}{The value of gamma at which to extract the solution.}

\item{supportSize}{The number of non-zeros each solution extracted will
contain. If no solutions have `supportSize` non-zeros, solutions with
the closest number will be extracted.}

\item{...}{ignore}
}
\description{
Extracts a specific solution in the regularization path.
}
\details{
If both lambda and gamma are not supplied, then a matrix of coefficients
for all the solutions in the regularization path is returned. If lambda is
supplied but gamma is not, the smallest value of gamma is used.
}
