% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare.networks}
\alias{compare.networks}
\title{compare.networks}
\usage{
compare.networks(
  weightDT,
  degreeValue = 0,
  weightValue = 0,
  legend.position = "bottom"
)
}
\arguments{
\item{weightDT}{weightDT}

\item{degreeValue}{degreeValue}

\item{weightValue}{weightValue}

\item{legend.position}{legend.position}
}
\value{
ggplot2 object
}
\description{
Ref: https://mp.weixin.qq.com/s/f3Hquw0m4ucGUieSjMyang
}
\examples{
library(inferCSN)
data("exampleMatrix")
weightDT <- inferCSN(exampleMatrix)
g <- compare.networks(weightDT[1:50, ])
g

}
