% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.gene.rank.R
\name{compute.gene.rank}
\alias{compute.gene.rank}
\title{compute.gene.rank}
\usage{
compute.gene.rank(weightDT, directedGraph = FALSE)
}
\arguments{
\item{weightDT}{The weight data table of network.}

\item{directedGraph}{If GRN is directed or not}
}
\value{
A data.table with three columns
}
\description{
compute.gene.rank
}
\details{
Function to compute page rank of TF+target networks
}
\examples{
data("exampleDataMatrix")
weightDT <- inferCSN(exampleDataMatrix)
ranks <- compute.gene.rank(weightDT)
head(ranks)
}
