% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossweighting.R
\name{cross.corr}
\alias{cross.corr}
\title{cross.corr}
\usage{
cross.corr(grn_row, expDat, lag)
}
\arguments{
\item{grn_row}{grn_row}

\item{expDat}{expDat}

\item{lag}{lag}
}
\value{
offset
}
\description{
cross.corr
}
