% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossweighting.R
\name{crossweight.params}
\alias{crossweight.params}
\title{crossweight.params
 estimates min and max values for crossweighting
 for now assumes uniform cell density across pseudotime/only considers early time
 this needs to be refined if it's to be useful...}
\usage{
crossweight.params(expDat, xdyn, pseudotime_min = 0.005, pseudotime_max = 0.01)
}
\arguments{
\item{expDat}{expDat}

\item{xdyn}{xdyn}

\item{pseudotime_min}{pseudotime_min}

\item{pseudotime_max}{pseudotime_max}
}
\description{
crossweight.params
 estimates min and max values for crossweighting
 for now assumes uniform cell density across pseudotime/only considers early time
 this needs to be refined if it's to be useful...
}
