% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.processing.R
\name{data.processing}
\alias{data.processing}
\title{data.processing}
\usage{
data.processing(data, normalize = normalize, verbose = verbose)
}
\arguments{
\item{data}{An object}

\item{normalize}{Data normalize}

\item{verbose}{Print detailed information}
}
\value{
Matrix
}
\description{
data.processing
}
