% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DT2Matrix}
\alias{DT2Matrix}
\title{Switch weight data table to weight matrix}
\usage{
DT2Matrix(weightDT = NULL)
}
\arguments{
\item{weightDT}{The weight data table of network.}
}
\value{
Weight matrix
}
\description{
Switch weight data table to weight matrix
}
