% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.genes.R
\name{dynamic.genes}
\alias{dynamic.genes}
\title{dynamic.genes
Find genes expressed dynamically
uses slingshot approach}
\usage{
dynamic.genes(matrix, matedata, path = NULL, method = "gam", matrixRaw = NULL)
}
\arguments{
\item{matrix}{properly normalized expression matrix}

\item{matedata}{sample table that includes pseudotime, rownames = cells, and a group column}

\item{path}{vector of group names to include}

\item{method}{method to find dynamic genes. Defaults to "gam", if "tradeseq", must provide matrixRaw}

\item{matrixRaw}{raw expression matrix, required if method="tradeseq"}

\item{group_column}{column name in matedata annotating groups in the path}

\item{pseudotime_column}{column name in matedata annotating pseudotime or latent time}
}
\value{
pvals and cell info
}
\description{
dynamic.genes
Find genes expressed dynamically
uses slingshot approach
}
