% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dynamic.networks}
\alias{dynamic.networks}
\title{dynamic.networks}
\usage{
dynamic.networks(weightList, regulators = NULL, legend.position = "right")
}
\arguments{
\item{weightList}{weightList}

\item{regulators}{regulators}

\item{legend.position}{legend.position}
}
\value{
A list of ggplot2 objects
}
\description{
Plot of dynamic networks
}
\examples{
data("exampleDataMatrix")
weightList <- inferCSN(exampleDataMatrix)
p <- dynamic.networks(weightList)
p
}
