% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleDataMatrix}
\alias{exampleDataMatrix}
\title{Test data}
\format{
A matrix with 5000 rows and 18 variables:
\describe{
  \item{matrix}{The matrix used for reconstruct gene regulatory network}
  ...
}
}
\source{
\url{NULL}
}
\description{
A dataset
diamonds.
}
\details{
Success!!
}
