% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_net.R
\name{extract_net}
\alias{extract_net}
\title{Create GRN table}
\usage{
extract_net(zscores, corrMatrix, genes, threshold, self_remove = TRUE)
}
\arguments{
\item{zscores}{zscores}

\item{corrMatrix}{zscores}

\item{genes}{zscores}

\item{threshold}{zscores}

\item{self_remove}{zscores}
}
\value{
data frame of GRN
}
\description{
Create GRN table
}
