% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.calculate.R
\name{figure.format}
\alias{figure.format}
\title{figure.format}
\usage{
figure.format(string)
}
\arguments{
\item{string}{A string of file name}
}
\value{
logic value
}
\description{
Check figure format
}
