% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filter_sort_matrix}
\alias{filter_sort_matrix}
\title{Filter and sort a network matrix}
\usage{
filter_sort_matrix(network_matrix, regulators = NULL, targets = NULL)
}
\arguments{
\item{network_matrix}{Network weight matrix.}

\item{regulators, targets}{Nodes to include.}
}
\value{
A filtered and sorted matrix.
}
\description{
Filter and sort a network matrix
}
