% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.genes.R
\name{gam.fit}
\alias{gam.fit}
\title{gam.fit}
\usage{
gam.fit(matrix, genes, celltime)
}
\arguments{
\item{matrix}{matrix}

\item{genes}{genes}

\item{celltime}{celltime}
}
\value{
list
}
\description{
gam.fit
}
