% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%s\%}
\alias{\%s\%}
\title{Value selection operator}
\usage{
a \%s\% b
}
\arguments{
\item{a}{The left side value to check}

\item{b}{The right side value to use if a is NULL}
}
\description{
This operator returns the left side if it's not NULL,
otherwise it returns the right side.
}
\examples{
NULL \%s\% 10
5 \%s\% 10
}
