% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN-package.R
\docType{package}
\name{inferCSN-package}
\alias{inferCSN-package}
\title{Inferring Cell-Specific Gene Regulatory Network}
\source{
\url{https://mengxu98.github.io/inferCSN/}
}
\description{
Inferring cell-type specific gene regulatory network using sparse regression model from single-cell transcriptomic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/inferCSN/}
  \item Report bugs at \url{https://github.com/mengxu98/inferCSN/issues}
}

}
\author{
Meng Xu (Maintainer), \email{mengxu98@qq.com}
}
