% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inferCSN.cvfit}
\alias{inferCSN.cvfit}
\title{Cross Validation}
\usage{
inferCSN.cvfit(
  x,
  y,
  loss = "SquaredError",
  penalty = "L0",
  algorithm = "CD",
  maxSuppSize = 100,
  nLambda = 100,
  nGamma = 10,
  gammaMax = 10,
  gammaMin = 1e-04,
  partialSort = TRUE,
  maxIters = 200,
  rtol = 1e-06,
  atol = 1e-09,
  activeSet = TRUE,
  activeSetNum = 3,
  maxSwaps = 100,
  scaleDownFactor = 0.8,
  screenSize = 1000,
  autoLambda = NULL,
  lambdaGrid = list(),
  nFolds = 10,
  seed = 1,
  excludeFirstK = 0,
  intercept = TRUE,
  lows = -Inf,
  highs = Inf
)
}
\arguments{
\item{x}{The data matrix}

\item{y}{The response vector}

\item{loss}{The loss function. Currently support the choices "SquaredError" (for regression), "Logistic" (for logistic regression), and "SquaredHinge" (for smooth SVM).}

\item{penalty}{The type of regularization}

\item{algorithm}{The type of algorithm used to minimize the objective function. Currently "CD" and "CDPSI" are supported.
"CD" is a variant of cyclic coordinate descent and runs very fast
"CDPSI" performs local combinatorial search on top of CD and typically achieves higher quality solutions (at the expense
of increased running time)}

\item{maxSuppSize}{The maximum support size at which to terminate the regularization path. Recommend setting
this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small
portion of non-zeros}

\item{nLambda}{The number of Lambda values to select}

\item{nGamma}{The number of Gamma values to select}

\item{gammaMax}{The maximum value of Gamma when using the L0L2 penalty.}

\item{gammaMin}{The minimum value of Gamma when using the L0L2 penalty.}

\item{partialSort}{If TRUE partial sorting will be used for sorting the coordinates to do greedy cycling. Otherwise, full sorting is used.}

\item{maxIters}{The maximum number of iterations (full cycles) for CD per grid point.}

\item{rtol}{The relative tolerance which decides when to terminate optimization (based on the relative change in the objective between iterations).}

\item{atol}{The absolute tolerance which decides when to terminate optimization (based on the absolute L2 norm of the residuals).}

\item{activeSet}{If TRUE, performs active set updates.}

\item{activeSetNum}{The number of consecutive times a support should appear before declaring support stabilization.}

\item{maxSwaps}{The maximum number of swaps used by CDPSI for each grid point.}

\item{scaleDownFactor}{This parameter decides how close the selected Lambda values are}

\item{screenSize}{The number of coordinates to cycle over when performing initial correlation screening}

\item{autoLambda}{Ignored parameter. Kept for backwards compatibility}

\item{lambdaGrid}{A grid of Lambda values to use in computing the regularization path}

\item{nFolds}{The number of folds for cross-validation.}

\item{seed}{The seed used in randomly shuffling the data for cross-validation}

\item{excludeFirstK}{This parameter takes non-negative integers}

\item{intercept}{If FALSE, no intercept term is included in the model.}

\item{lows}{Lower bounds for coefficients}

\item{highs}{Upper bounds for coefficients}
}
\value{
An S3 object of type "inferCSNCV" describing the regularization path
}
\description{
Computes a regularization path and performs K-fold cross-validation
}
