% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.fit.R
\name{inferCSN.fit}
\alias{inferCSN.fit}
\title{inferCSN.fit}
\usage{
inferCSN.fit(
  X,
  Y,
  penalty = penalty,
  nFolds = 10,
  seed = 1,
  maxSuppSize = maxSuppSize,
  nGamma = 5,
  gammaMin = 1e-04,
  gammaMax = 10
)
}
\arguments{
\item{X}{The rows are samples and the columns are genes of the matrix}

\item{gammaMax}{}
}
\description{
inferCSN.fit
}
