% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.plot.R
\name{inferCSN.plot.dynamic.networks}
\alias{inferCSN.plot.dynamic.networks}
\title{Plot of dynamic networks}
\usage{
inferCSN.plot.dynamic.networks(
  weightList,
  regulators = NULL,
  onlyRegulators = TRUE,
  order = NULL,
  thresh = NULL,
  legend.position = "right"
)
}
\arguments{
\item{weightList}{weightList}

\item{regulators}{regulators}

\item{onlyRegulators}{onlyregulators}

\item{order}{order = NULL}

\item{thresh}{thresh = NULL}

\item{legend.position}{legend.position}
}
\value{
A list of ggplot2 objects
}
\description{
Plot
}
