% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.plot.R
\name{inferCSN.plot}
\alias{inferCSN.plot}
\title{inferCSN.plot}
\usage{
inferCSN.plot(data, plotType = NULL)
}
\arguments{
\item{data}{A long data table}

\item{plotType}{}
}
\value{
A ggplot2 object
}
\description{
inferCSN.plot
}
