% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN-package.R
\name{infercsn_logo}
\alias{infercsn_logo}
\title{The logo of inferCSN}
\usage{
infercsn_logo(unicode = cli::is_utf8_output())
}
\arguments{
\item{unicode}{Unicode symbols on UTF-8 platforms.
Default is \link[cli:is_utf8_output]{cli::is_utf8_output}.}
}
\value{
A character vector.
}
\description{
The inferCSN logo, using ASCII or Unicode characters
Use \link[cli:ansi_strip]{cli::ansi_strip} to get rid of the colors.
}
\examples{
infercsn_logo()
}
\references{
\url{https://github.com/tidyverse/tidyverse/blob/main/R/logo.R}
}
