% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{net.format}
\alias{net.format}
\title{net.format}
\usage{
net.format(weightList, regulators = NULL)
}
\arguments{
\item{weightList}{weightList of CSN}

\item{regulators}{Regulators list}
}
\value{
Format weight table
}
\description{
Format weight table
}
