% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{network.heatmap}
\alias{network.heatmap}
\title{network.heatmap}
\usage{
network.heatmap(
  weightDT = NULL,
  heatmapSize = NULL,
  heatmapTitle = NULL,
  heatmapColor = NULL,
  showNames = FALSE
)
}
\arguments{
\item{weightDT}{The weight data table of network.}

\item{heatmapSize}{heatmapSize}

\item{heatmapTitle}{heatmapTitle}

\item{heatmapColor}{heatmapColor}

\item{showNames}{showNames}
}
\value{
heatmap
}
\description{
network.heatmap
}
\examples{
data("exampleMatrix")
data("exampleGroundTruth")
weightDT <- inferCSN(exampleMatrix)
p1 <- network.heatmap(exampleGroundTruth,
                      heatmapTitle = "Ground truth")
p1
p2 <- network.heatmap(weightDT,
                      heatmapTitle = "inferCSN")
p2
p3 <- network.heatmap(weightDT,
                      heatmapTitle = "inferCSN",
                      heatmapColor = c("#20a485", "#410054", "#fee81f"))
p3
p4 <- network.heatmap(weightDT,
                      heatmapTitle = "inferCSN",
                      heatmapColor = c("#20a485", "white", "#fee81f"))
p4
p5 <- network.heatmap(weightDT,
                      heatmapTitle = "inferCSN",
                      showNames = TRUE)
p5
}
