% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{network_format}
\alias{network_format}
\title{Format a network table}
\usage{
network_format(
  network_table,
  regulators = NULL,
  targets = NULL,
  abs_weight = TRUE
)
}
\arguments{
\item{network_table}{Network edge table.}

\item{regulators, targets}{Nodes to include.}

\item{abs_weight}{Whether to use absolute weights and add interaction signs.}
}
\value{
A formatted network edge table.
}
\description{
Format a network table
}
