% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pearson_correlation}
\alias{pearson_correlation}
\title{Correlation and covariance calculation for sparse matrix}
\usage{
pearson_correlation(x, y = NULL)
}
\arguments{
\item{x}{Sparse matrix or character vector.}

\item{y}{Sparse matrix or character vector.}
}
\description{
Correlation and covariance calculation for sparse matrix
}
