% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.inferCSN}
\alias{plot.inferCSN}
\title{Plot Regularization Path}
\usage{
\method{plot}{inferCSN}(x, gamma = 0, showLines = FALSE, ...)
}
\arguments{
\item{x}{The output of inferCSN.fit}

\item{gamma}{The value of gamma at which to plot.}

\item{showLines}{If TRUE, the lines connecting the points in the plot are shown.}

\item{...}{ignore}
}
\description{
Plots the regularization path for a given gamma.
}
