% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.inferCSNCV}
\alias{plot.inferCSNCV}
\title{Plot Cross-validation Errors}
\usage{
\method{plot}{inferCSNCV}(x, gamma = 0, ...)
}
\arguments{
\item{x}{The output of inferCSN.cvfit}

\item{gamma}{The value of gamma at which to plot.}

\item{...}{ignore}
}
\description{
Plots cross-validation errors for a given gamma.
}
