% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_dynamic_networks}
\alias{plot_dynamic_networks}
\title{Plot dynamic networks}
\usage{
plot_dynamic_networks(
  network_table,
  celltypes_order,
  ntop = 10,
  title = NULL,
  theme_type = "theme_void",
  plot_type = "ggplot",
  layout = "fruchtermanreingold",
  nrow = 2,
  figure_save = FALSE,
  figure_name = NULL,
  figure_width = 6,
  figure_height = 6,
  seed = 1
)
}
\arguments{
\item{network_table}{Network edge table.}

\item{celltypes_order}{Cell-type order.}

\item{ntop}{Number of top genes to plot.}

\item{title}{Figure title.}

\item{theme_type}{Plot theme.}

\item{plot_type}{Output type.}

\item{layout}{Network layout.}

\item{nrow}{Number of rows.}

\item{figure_save}{Whether to save the figure.}

\item{figure_name}{Output filename.}

\item{figure_width, figure_height}{Figure dimensions.}

\item{seed}{Random seed.}
}
\value{
A dynamic network plot.
}
\description{
Plot dynamic networks
}
