% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot an embedding}
\usage{
plot_embedding(
  matrix,
  labels = NULL,
  method = "pca",
  colors = RColorBrewer::brewer.pal(length(unique(labels)), "Set1"),
  seed = 1,
  point_size = 1,
  cores = 1
)
}
\arguments{
\item{matrix}{Input matrix.}

\item{labels}{Point labels.}

\item{method}{Embedding method.}

\item{colors}{Point colors.}

\item{seed}{Random seed.}

\item{point_size}{Point size.}

\item{cores}{Number of threads.}
}
\value{
An embedding plot.
}
\description{
Plot an embedding
}
