% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot expression data}
\usage{
plot_scatter(
  data,
  smoothing_method = "lm",
  group_colors = RColorBrewer::brewer.pal(9, "Set1"),
  title_color = "black",
  title = NULL,
  col_title = NULL,
  row_title = NULL,
  legend_title = NULL,
  legend_position = "bottom",
  margins = "both",
  marginal_type = NULL,
  margins_size = 10,
  compute_correlation = TRUE,
  compute_correlation_method = "pearson",
  keep_aspect_ratio = TRUE,
  facet = FALSE,
  se = FALSE,
  pointdensity = TRUE
)
}
\arguments{
\item{data}{Input data.}

\item{smoothing_method}{Method for smoothing curve, `lm` or `loess`.}

\item{group_colors, title_color}{Group and title colors.}

\item{title, col_title, row_title, legend_title}{Plot titles.}

\item{legend_position}{The position of legend.}

\item{margins, marginal_type, margins_size}{Marginal plot controls.}

\item{compute_correlation, compute_correlation_method}{Correlation controls.}

\item{keep_aspect_ratio}{Whether to use a 1:1 aspect ratio.}

\item{facet}{Whether to facet by group.}

\item{se}{Whether to show smoothing uncertainty.}

\item{pointdensity}{Whether to show point density.}
}
\value{
A ggplot object.
}
\description{
Plot expression data
}
