% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_static_networks}
\alias{plot_static_networks}
\title{Plot dynamic networks}
\usage{
plot_static_networks(
  network_table,
  regulators = NULL,
  targets = NULL,
  legend_position = "right"
)
}
\arguments{
\item{network_table}{Network edge table.}

\item{regulators, targets}{Nodes to include.}

\item{legend_position}{The position of legend.}
}
\value{
A ggplot2 object
}
\description{
Plot dynamic networks
}
\examples{
data(example_matrix)
network_table <- inferCSN(example_matrix)
example_edge <- network_table[1, ]
plot_static_networks(
  network_table,
  regulators = example_edge$regulator
)
plot_static_networks(
  network_table,
  targets = example_edge$target
)
plot_static_networks(
  network_table,
  regulators = example_edge$regulator,
  targets = example_edge$target
)
}
