% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_weight_distribution}
\alias{plot_weight_distribution}
\title{plot_weight_distribution}
\usage{
plot_weight_distribution(net_table)
}
\arguments{
\item{net_table}{Input data frame.}
}
\value{
ggplot object
}
\description{
plot_weight_distribution
}
\examples{
data("example_matrix")
net_table <- inferCSN(example_matrix)
plot_weight_distribution(net_table)
}
