% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.srm}
\alias{predict.srm}
\alias{predict.srm_cv}
\title{Predicts response for a given sample}
\usage{
\method{predict}{srm}(object, newx, lambda = NULL, gamma = NULL, ...)

\method{predict}{srm_cv}(object, newx, lambda = NULL, gamma = NULL, ...)
}
\arguments{
\item{object}{The output of sparse_regression.}

\item{newx}{A matrix on which predictions are made. The matrix should have p columns}

\item{lambda}{The value of lambda to use for prediction.
A summary of the lambdas in the regularization path can be obtained using \code{\link{print.srm}}.}

\item{gamma}{The value of gamma to use for prediction.
A summary of the gammas in the regularization path can be obtained using \code{\link{print.srm}}.}

\item{...}{Other parameters}
}
\value{
Return predict value
}
\description{
Predicts response for a given sample
}
\details{
If both lambda and gamma are not supplied,
then a matrix of predictions for all the solutions in the regularization path is returned.
If lambda is supplied but gamma is not, the smallest value of gamma is used.
In case of logistic regression, probability values are returned.
}
