% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.calculate.R
\name{prepare.performance.data}
\alias{prepare.performance.data}
\title{prepare.performance.data}
\usage{
prepare.performance.data(weight_table, ground_truth)
}
\arguments{
\item{weight_table}{The weight data table of network}

\item{ground_truth}{Ground truth for calculate AUC}
}
\value{
Formated data
}
\description{
prepare.performance.data
}
