% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_evaluation.R
\name{prepare_network_data}
\alias{prepare_network_data}
\title{Prepare network data}
\usage{
prepare_network_data(network_table, ground_truth)
}
\arguments{
\item{network_table}{The weight data table of network}

\item{ground_truth}{Ground truth for calculate AUC}
}
\value{
Formated network table
}
\description{
Formated network table
}
