% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.inferCSN}
\alias{print.inferCSN}
\alias{print.inferCSNCV}
\title{Print inferCSN.fit object}
\usage{
\method{print}{inferCSN}(x, ...)

\method{print}{inferCSNCV}(x, ...)
}
\arguments{
\item{x}{The output of inferCSN.fit or inferCSN.cvfit}

\item{...}{ignore}
}
\description{
Prints a summary of inferCSN.fit
}
