% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.inferCSNObject}
\alias{print.inferCSNObject}
\alias{print.inferCSNCVObject}
\title{Print inferCSN.fit object}
\usage{
\method{print}{inferCSNObject}(x, ...)

\method{print}{inferCSNCVObject}(x, ...)
}
\arguments{
\item{x}{The output of inferCSN.fit or inferCSN.cvfit}

\item{...}{ignore}
}
\description{
Prints a summary of inferCSN.fit
}
