% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.SRM_fit}
\alias{print.SRM_fit}
\alias{print.SRM_fit_CV}
\title{Prints a summary of fit_sparse_regression}
\usage{
\method{print}{SRM_fit}(x, ...)

\method{print}{SRM_fit_CV}(x, ...)
}
\arguments{
\item{x}{The output of fit_sparse_regression or inferCSN.cvfit}

\item{...}{Other parameters}
}
\value{
Return information of fit_sparse_regression

Return information of fit_sparse_regression
}
\description{
Prints a summary of fit_sparse_regression
}
