% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossweighting.R
\name{score.offset}
\alias{score.offset}
\title{score.offset}
\usage{
score.offset(score, offset, min = 2, max = 20, symmetric_filter = FALSE)
}
\arguments{
\item{score}{score}

\item{offset}{offset}

\item{min}{min}

\item{max}{max}

\item{symmetric_filter}{symmetric_filter}
}
\value{
score.offset
}
\description{
score.offset
}
