% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossweighting.R
\name{score_offset}
\alias{score_offset}
\title{score_offset}
\usage{
score_offset(score, offset, min = 2, max = 20, symmetric_filter = FALSE)
}
\arguments{
\item{score}{score}

\item{offset}{offset}

\item{min}{min}

\item{max}{max}

\item{symmetric_filter}{symmetric_filter}
}
\value{
score_offset
}
\description{
score_offset
}
