% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sparse.regression}
\alias{sparse.regression}
\title{Sparse regression model}
\usage{
sparse.regression(
  X,
  y,
  crossValidation = FALSE,
  penalty = "L0",
  algorithm = "CD",
  maxSuppSize = NULL,
  nFolds = 10,
  verbose = FALSE
)
}
\arguments{
\item{X}{The data matrix}

\item{y}{The response vector}

\item{crossValidation}{Check whether cross validation is used}

\item{penalty}{The type of regularization.
This can take either one of the following choices: "L0"and "L0L2".
For high-dimensional and sparse data, such as single-cell transcriptome data, "L0L2" is more effective}

\item{algorithm}{Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time}

\item{maxSuppSize}{The number of non-zore coef, this value will affect the final performance}

\item{nFolds}{The number of folds for cross-validation}

\item{verbose}{Print detailed information}
}
\value{
The coefficients
}
\description{
Sparse regression model
}
