% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse_cov_cor}
\alias{sparse_cov_cor}
\title{Fast correlation and covariance calcualtion for sparse matrices}
\usage{
sparse_cov_cor(x, y = NULL)
}
\arguments{
\item{x}{Sparse matrix or character vector.}

\item{y}{Sparse matrix or character vector.}
}
\description{
Fast correlation and covariance calcualtion for sparse matrices
}
