% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_indices}
\alias{split_indices}
\title{Split indices.}
\usage{
split_indices(group, n = 0L)
}
\arguments{
\item{group}{Integer indices}

\item{n}{The largest integer (may not appear in index).
This is hint: if the largest value of \code{group} is bigger than \code{n},
the output will silently expand.}
}
\value{
A list of vectors of indices.
}
\description{
An optimised version of split for the special case of splitting row indices into groups.
}
\examples{
split_indices(sample(10, 100, rep = TRUE))
split_indices(sample(10, 100, rep = TRUE), 10)
}
\references{
https://github.com/hadley/plyr/blob/d57f9377eb5d56107ba3136775f2f0f005f33aa3/src/split-numeric.cpp#L20
}
