% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.regression.R
\name{sub.inferCSN}
\alias{sub.inferCSN}
\title{Sparse regression model for single gene}
\usage{
sub.inferCSN(
  regulatorsMatrix,
  targetsMatrix,
  target = NULL,
  crossValidation = FALSE,
  penalty = "L0",
  algorithm = "CD",
  maxSuppSize = NULL,
  nFolds = 10,
  kFolds = NULL,
  rThreshold = 0,
  verbose = FALSE
)
}
\arguments{
\item{regulatorsMatrix}{regulatorsMatrix}

\item{targetsMatrix}{targetsMatrix}

\item{target}{target}

\item{crossValidation}{Check whether cross validation is used}

\item{penalty}{The type of regularization.
This can take either one of the following choices: "L0"and "L0L2".
For high-dimensional and sparse data, such as single-cell transcriptome data, "L0L2" is more effective}

\item{algorithm}{Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time}

\item{maxSuppSize}{The number of non-zore coef, this value will affect the final performance}

\item{nFolds}{The number of folds for cross-validation}

\item{kFolds}{The number of folds for sample split}

\item{rThreshold}{rThreshold}

\item{verbose}{Print detailed information}
}
\value{
The weight data table of sub-network
}
\description{
Sparse regression model for single gene
}
