% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{subsampling}
\alias{subsampling}
\title{Subsample an expression matrix}
\usage{
subsampling(
  matrix,
  subsampling_method = c("sample", "meta_cells", "pseudobulk"),
  subsampling_ratio = 1,
  seed = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{matrix}{Input matrix.}

\item{subsampling_method}{Subsampling strategy.}

\item{subsampling_ratio}{Fraction retained or aggregated.}

\item{seed}{Random seed.}

\item{verbose}{Whether to report progress.}

\item{...}{Additional arguments.}
}
\value{
The subsampled matrix.
}
\description{
Subsample an expression matrix
}
