% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_to_matrix}
\alias{table_to_matrix}
\title{Switch weight table}
\usage{
table_to_matrix(weight_table)
}
\arguments{
\item{weight_table}{The weight data table of network}
}
\value{
Format weight matrix
}
\description{
Switch weight table
}
