% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_filter.R
\name{weight_filter}
\alias{weight_filter}
\title{weight_filter}
\usage{
weight_filter(network_table, method = "max")
}
\arguments{
\item{network_table}{network_table}

\item{method}{method}
}
\value{
Filtered network table
}
\description{
weight_filter
}
\examples{
data("example_matrix")
data("example_ground_truth")
network_table <- inferCSN(example_matrix)
network_table_new <- weight_filter(network_table)
network.heatmap(
  example_ground_truth[, 1:3],
  heatmap_title = "Ground truth",
  show_names = TRUE,
  rect_color = "gray90"
)
network.heatmap(
  network_table,
  heatmap_title = "Raw",
  show_names = TRUE,
  rect_color = "gray90"
)
network.heatmap(
  network_table_new,
  heatmap_title = "Filtered",
  show_names = TRUE,
  rect_color = "gray90"
)

auc.calculate(
  network_table,
  example_ground_truth,
  plot = TRUE
)
auc.calculate(
  network_table_new,
  example_ground_truth,
  plot = TRUE
)
}
