% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weight_sift}
\alias{weight_sift}
\title{Weight sift}
\usage{
weight_sift(table)
}
\arguments{
\item{table}{A data frame with three columns: "regulator", "target", and "weight".}
}
\description{
Remove edges with smaller weights in the reverse direction.
}
\examples{
data("example_matrix")
network_table <- inferCSN(example_matrix)
weight_sift(network_table) |> head()
}
