% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_matrix}
\alias{as_matrix}
\title{Attempts to turn a dgCMatrix into a dense matrix}
\usage{
as_matrix(x)
}
\arguments{
\item{x}{A matrix.}
}
\description{
Attempts to turn a dgCMatrix into a dense matrix
}
\examples{
sparse_matrix <- Matrix::sparseMatrix(
  i = sample(1:200, 50),
  j = sample(1:200, 50),
  x = rnorm(50),
  dims = c(200, 200),
  dimnames = list(
    paste0("a", rep(1:200)),
    paste0("b", rep(1:200))
  )
)

identical(
  as.matrix(sparse_matrix),
  as_matrix(sparse_matrix)
)
}
