% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.calculate.R
\name{auc.calculate}
\alias{auc.calculate}
\title{auc.calculate}
\usage{
auc.calculate(
  weightDT = NULL,
  groundTruth = NULL,
  plot = FALSE,
  fileSave = NULL,
  interaction = FALSE
)
}
\arguments{
\item{weightDT}{The weight data table of network.}

\item{groundTruth}{Ground truth for calculate AUC.}

\item{plot}{If true, draw and print figure of AUC.}

\item{fileSave}{The figure name}

\item{interaction}{If true, consider the positivity and negativity of interaction.}
}
\value{
AUC values and figure
}
\description{
AUC value calculate
}
\examples{
data("exampleDataMatrix")
data("exampleDataGroundTruth")
weightDT <- inferCSN(exampleDataMatrix, cores = 1, verbose = TRUE, algorithm = "CDPSI")
auc <- auc.calculate(weightDT, exampleDataGroundTruth, plot = TRUE)
head(auc)
}
