% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.calculate.R
\name{auc.calculate}
\alias{auc.calculate}
\title{auc.calculate
 AUC value calculate}
\usage{
auc.calculate(
  weightList = NULL,
  goldRef = NULL,
  goldPathway = NULL,
  plot = FALSE
)
}
\arguments{
\item{weightList}{weightList}

\item{goldRef}{goldRef}

\item{goldPathway}{goldPathway}

\item{plot}{plot}
}
\value{
AUC
}
\description{
auc.calculate
 AUC value calculate
}
\examples{
   data("exampleDataMatrix")
   data("exampleDataGroundTruth")
   weightList <- inferCSN(exampleDataMatrix, cores = 2)
   auc <- auc.calculate(weightList, exampleDataGroundTruth)
}
