% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.calculate.R
\name{auc.calculate}
\alias{auc.calculate}
\title{AUC value calculate}
\usage{
auc.calculate(
  weightDT,
  groundTruth,
  plot = FALSE,
  fileSave = NULL,
  interaction = FALSE
)
}
\arguments{
\item{weightDT}{The weight data table of network}

\item{groundTruth}{Ground truth for calculate AUC}

\item{plot}{If true, draw and print figure of AUC}

\item{fileSave}{The figure name}

\item{interaction}{If true, consider the positivity or negativity of interaction}
}
\value{
AUC values and figure
}
\description{
AUC value calculate
}
\examples{
library(inferCSN)
data("exampleMatrix")
data("exampleGroundTruth")
weightDT <- inferCSN(exampleMatrix)
auc <- auc.calculate(weightDT, exampleGroundTruth, plot = TRUE)
head(auc)

}
