% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.calculate.R
\name{auc.calculate}
\alias{auc.calculate}
\title{AUC value calculate}
\usage{
auc.calculate(
  network_table,
  ground_truth,
  plot = FALSE,
  line_color = "#1563cc",
  line_width = 1
)
}
\arguments{
\item{network_table}{The weight data table of network}

\item{ground_truth}{Ground truth for calculate AUC}

\item{plot}{If true, draw and print figure of AUC}

\item{line_color}{The color of line in the figure}

\item{line_width}{The width of line in the figure}
}
\value{
AUC values and figure
}
\description{
AUC value calculate
}
\examples{
data("example_matrix")
data("example_ground_truth")
network_table <- inferCSN(example_matrix)
auc.calculate(network_table, example_ground_truth, plot = TRUE)
}
