% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.gene.rank.R
\name{calculate.gene.rank}
\alias{calculate.gene.rank}
\title{Calculate and rank TFs in network}
\usage{
calculate.gene.rank(weight_table, directed = FALSE)
}
\arguments{
\item{weight_table}{The weight data table of network}

\item{directed}{If GRN is directed or not}
}
\value{
A data.table with three columns
}
\description{
Calculate and rank TFs in network
}
\examples{
library(inferCSN)
data("example_matrix")
weight_table <- inferCSN(example_matrix)
ranks <- calculate.gene.rank(weight_table)
head(ranks)
}
