% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_evaluation.R
\name{calculate_ji}
\alias{calculate_ji}
\title{Calculate Jaccard Index}
\usage{
calculate_ji(network_table, ground_truth)
}
\arguments{
\item{network_table}{A data frame of predicted network structure}

\item{ground_truth}{A data frame of ground truth network}
}
\value{
A list containing the metric
}
\description{
Calculates Jaccard Index (JI) metric
}
\examples{
data("example_matrix")
data("example_ground_truth")
network_table <- inferCSN(example_matrix)
calculate_ji(
  network_table,
  example_ground_truth
)
}
