% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_evaluation.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate metrics}
\usage{
calculate_metrics(
  network_table,
  ground_truth,
  type = "all",
  plot = FALSE,
  line_color = "#1563cc",
  line_width = 1
)
}
\arguments{
\item{network_table}{The weight data table of network}

\item{ground_truth}{Ground truth for calculate AUC}

\item{type}{The type of metrics to calculate}

\item{plot}{If true, draw and print figure of AUC}

\item{line_color}{The color of line in the figure}

\item{line_width}{The width of line in the figure}
}
\value{
AUC values and figure
}
\description{
Calculate metrics
}
\examples{
data("example_matrix")
data("example_ground_truth")
inferCSN(example_matrix) |>
  calculate_metrics(example_ground_truth, plot = TRUE)
}
