% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_evaluation.R
\name{calculate_recall}
\alias{calculate_recall}
\title{Calculate Recall Metric}
\usage{
calculate_recall(network_table, ground_truth)
}
\arguments{
\item{network_table}{A data frame of predicted network structure}

\item{ground_truth}{A data frame of ground truth network}
}
\value{
A list containing the metric
}
\description{
Calculates recall metric
}
\examples{
data("example_matrix")
data("example_ground_truth")
network_table <- inferCSN(example_matrix)
calculate_recall(
  network_table,
  example_ground_truth
)
}
