% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.inferCSN}
\alias{coef.inferCSN}
\alias{coef.inferCSNCV}
\title{Extract Solutions}
\usage{
\method{coef}{inferCSN}(object, lambda = NULL, gamma = NULL, supportSize = NULL, ...)

\method{coef}{inferCSNCV}(object, lambda = NULL, gamma = NULL, ...)
}
\arguments{
\item{object}{The output of inferCSN.fit or inferCSN.cvfit}

\item{lambda}{The value of lambda at which to extract the solution.}

\item{gamma}{The value of gamma at which to extract the solution.}

\item{supportSize}{The number of non-zeros each solution extracted will
contain. If no solutions have `supportSize` non-zeros, solutions with
the closest number will be extracted.}

\item{...}{ignore}
}
\description{
Extracts a specific solution in the regularization path.
}
\details{
If both lambda and gamma are not supplied, then a matrix of coefficients
for all the solutions in the regularization path is returned. If lambda is
supplied but gamma is not, the smallest value of gamma is used.
}
\examples{
# Generate synthetic data for this example
data <- GenSynthetic(n=500,p=1000,k=10,seed=1)
X = data$X
y = data$y

# Fit an L0L2 Model with 10 values of Gamma ranging from 0.0001 to 10,
# using coordinate descent
fit <- inferCSN.fit(X, y, penalty="L0L2", maxSuppSize=50,
                   nGamma=10, gammaMin=0.0001, gammaMax = 10)
print(fit)
# Extract the coefficients of the solution at lambda = 0.0361829
# and gamma = 0.0001
coef(fit, lambda=0.0361829, gamma=0.0001)
# Extract the coefficients of all the solutions in the path
coef(fit)
# Extract the coefficients of the solution at supportSize = 10
# and gamma = 0.0001
coef(fit, supportSize=10, gamma=0.0001)
}
