% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.inferCSN}
\alias{coef.inferCSN}
\alias{coef.inferCSNCV}
\title{Extracts a specific solution in the regularization path}
\usage{
\method{coef}{inferCSN}(object, lambda = NULL, gamma = NULL, supportSize = NULL, ...)

\method{coef}{inferCSNCV}(object, lambda = NULL, gamma = NULL, ...)
}
\arguments{
\item{object}{The output of inferCSN.fit or inferCSN.cvfit}

\item{lambda}{The value of lambda at which to extract the solution}

\item{gamma}{The value of gamma at which to extract the solution}

\item{supportSize}{The number of non-zeros each solution extracted will contain}

\item{...}{Other parameters}
}
\value{
Return the specific solution

Return the specific solution
}
\description{
Extracts a specific solution in the regularization path
}
