% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.gene.rank.R
\name{compute.gene.rank}
\alias{compute.gene.rank}
\title{Compute and rank TFs in network}
\usage{
compute.gene.rank(weightDT, directedGraph = FALSE)
}
\arguments{
\item{weightDT}{The weight data table of network}

\item{directedGraph}{If GRN is directed or not}
}
\value{
A data.table with three columns
}
\description{
Compute and rank TFs in network
}
\examples{
library(inferCSN)
data("exampleMatrix")
weightDT <- inferCSN(exampleMatrix)
ranks <- compute.gene.rank(weightDT)
head(ranks)

}
