% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.visualization.R
\name{contrast.networks}
\alias{contrast.networks}
\title{contrast.networks}
\usage{
contrast.networks(
  weight_table,
  degree_value = 0,
  weight_value = 0,
  legend.position = "bottom"
)
}
\arguments{
\item{weight_table}{weight_table}

\item{degree_value}{degree_value}

\item{weight_value}{weight_value}

\item{legend.position}{legend.position}
}
\value{
Return a ggplot2 object
}
\description{
contrast.networks
}
\examples{
library(inferCSN)
data("example_matrix")
weight_table <- inferCSN(example_matrix)
g <- contrast.networks(weight_table[1:50, ])
g
}
