% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossweight.R
\name{crossweight_params}
\alias{crossweight_params}
\title{estimates min and max values for crossweighting for now assumes uniform cell density across
pseudotime/only considers early time this needs to be refined if it's to be useful...}
\usage{
crossweight_params(
  matrix,
  meta_data,
  pseudotime_min = 0.005,
  pseudotime_max = 0.01
)
}
\arguments{
\item{matrix}{matrix}

\item{meta_data}{meta_data}

\item{pseudotime_min}{pseudotime_min}

\item{pseudotime_max}{pseudotime_max}
}
\value{
Params list
}
\description{
estimates min and max values for crossweighting for now assumes uniform cell density across
pseudotime/only considers early time this needs to be refined if it's to be useful...
}
