% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.processing.R
\name{data.ksmooth}
\alias{data.ksmooth}
\title{smooths expression
smooths expression across metadata in path}
\usage{
data.ksmooth(matrix, metadata, bandwidth = 0.25)
}
\arguments{
\item{matrix}{expression matrix}

\item{metadata}{sample table that includes "cell"  "pseudotime" "group"}

\item{bandwith}{bandwith for kernel smoother}
}
\value{
smoothed matrix
}
\description{
smooths expression
smooths expression across metadata in path
}
