% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.processing.R
\name{data.processing}
\alias{data.processing}
\title{data.processing}
\usage{
data.processing(data, normalize = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{A matrix, data table, Seurat or SingleCellExperiment object.}

\item{normalize}{[Default = FALSE] Data normalize.}

\item{verbose}{[Default = FALSE] Print detailed information.}
}
\value{
Normalized matrix
}
\description{
data.processing
}
