% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.genes.R
\name{dynamic.genes}
\alias{dynamic.genes}
\title{dynamic.genes}
\usage{
dynamic.genes(matrix, metadata, path = NULL, method = "gam", matrixRaw = NULL)
}
\arguments{
\item{matrix}{properly normalized expression matrix}

\item{metadata}{sample table that includes pseudotime, rownames = cells, and a group column}

\item{path}{vector of group names to include}

\item{method}{method to find dynamic genes. Defaults to "gam", if "tradeseq", must provide matrixRaw}

\item{matrixRaw}{raw expression matrix, required if method="tradeseq"}
}
\value{
pvals and cell info
}
\description{
Find genes expressed dynamically using slingshot approach
}
