% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dynamic.networks}
\alias{dynamic.networks}
\title{dynamic.networks}
\usage{
dynamic.networks(weightDT, regulators = NULL, legend.position = "right")
}
\arguments{
\item{weightDT}{weightDT}

\item{regulators}{regulators}

\item{legend.position}{legend.position}
}
\value{
A list of ggplot2 objects
}
\description{
Plot of dynamic networks
}
\examples{
data("exampleDataMatrix")
weightDT <- inferCSN(exampleDataMatrix)
p <- dynamic.networks(weightDT)
p
}
