% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dynamic.networks}
\alias{dynamic.networks}
\title{dynamic.networks}
\usage{
dynamic.networks(
  weightList,
  regulators = NULL,
  onlyRegulators = TRUE,
  order = NULL,
  thresh = NULL,
  legend.position = "right"
)
}
\arguments{
\item{weightList}{weightList}

\item{regulators}{[Default = NULL] regulators}

\item{onlyRegulators}{[Default = NULL] onlyregulators}

\item{order}{[Default = NULL] order = NULL}

\item{thresh}{[Default = NULL] thresh = NULL}

\item{legend.position}{[Default = "right"] legend.position}
}
\value{
A list of ggplot2 objects
}
\description{
Plot of dynamic networks
}
\examples{
data("exampleDataMatrix")
weightList <- inferCSN(exampleDataMatrix)
p <- dynamic.networks(weightList)
p
}
