% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.visualization.R
\name{dynamic.networks}
\alias{dynamic.networks}
\title{Plot of dynamic networks}
\usage{
dynamic.networks(weight_table, regulators = NULL, legend.position = "right")
}
\arguments{
\item{weight_table}{weight_table}

\item{regulators}{regulators}

\item{legend.position}{legend.position}
}
\value{
A list of ggplot2 objects
}
\description{
Plot of dynamic networks
}
\examples{
\dontrun{
# because `ggnetwork` package may be cause error
# you can install `igraph 1.6.0` or lower to deal error
# or waitting a new version of `ggnetwork` package
library(inferCSN)
data("example_matrix")
weight_table <- inferCSN(example_matrix)
g <- dynamic.networks(weight_table, regulators = weight_table[1, 1])
g
}
}
