% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.visualization.R
\name{dynamic.networks}
\alias{dynamic.networks}
\title{Plot of dynamic networks}
\usage{
dynamic.networks(weightDT, regulators = NULL, legend.position = "right")
}
\arguments{
\item{weightDT}{weightDT}

\item{regulators}{regulators}

\item{legend.position}{legend.position}
}
\value{
A list of ggplot2 objects
}
\description{
Plot of dynamic networks
}
\examples{
library(inferCSN)
data("exampleMatrix")
weightDT <- inferCSN(exampleMatrix)
g <- dynamic.networks(weightDT, regulators = weightDT[1, 1])
g

}
