% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{figure.format}
\alias{figure.format}
\title{Checking figure format}
\usage{
figure.format(string)
}
\arguments{
\item{string}{A string of file name}
}
\value{
Logic value
}
\description{
Checking figure format
}
