% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_sort_matrix}
\alias{filter_sort_matrix}
\title{Filter and sort matrix}
\usage{
filter_sort_matrix(weight_matrix, regulators = NULL, targets = NULL)
}
\arguments{
\item{weight_matrix}{The matrix of network weight.}

\item{regulators}{Regulators list.}

\item{targets}{Targets list.}
}
\value{
Filtered and sorted matrix
}
\description{
Filter and sort matrix
}
\examples{
data("example_matrix")
network_table <- inferCSN(example_matrix)
weight_matrix <- table_to_matrix(network_table)
filter_sort_matrix(weight_matrix)[1:6, 1:6]

filter_sort_matrix(
  weight_matrix,
  regulators = c("g1", "g2"),
  targets = c("g3", "g4")
)
}
