% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN-package.R
\docType{package}
\name{inferCSN-package}
\alias{inferCSN-package}
\title{\emph{\strong{inferCSN}}: \strong{infer}ring \strong{C}ell-\strong{S}pecific gene regulatory \strong{N}etwork}
\source{
\url{https://github.com/mengxu98/inferCSN}
}
\description{
An R package for \strong{infer}ring \strong{C}ell-\strong{S}pecific gene regulatory \strong{N}etwork from single-cell RNA data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/inferCSN/}
  \item Report bugs at \url{https://github.com/mengxu98/inferCSN/issues}
}

}
\author{
Meng xu (Maintainer), \email{mengxu98@qq.com}
}
