% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN-package.R
\docType{package}
\name{inferCSN-package}
\alias{inferCSN-package}
\title{inferCSN: Inferring Cell-Specific Gene Regulatory Network}
\source{
\url{https://github.com/mengxu98/inferCSN}
}
\description{
A method for inferring cell-specific gene regulatory network from single-cell sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/inferCSN/}
  \item Report bugs at \url{https://github.com/mengxu98/inferCSN/issues}
}

}
\author{
Meng xu (Maintainer), \email{mengxu98@qq.com}
}
