% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.fit.R
\name{inferCSN.fit}
\alias{inferCSN.fit}
\title{inferCSN.fit}
\usage{
inferCSN.fit(
  X,
  Y,
  penalty = penalty,
  nFolds = 10,
  seed = 1,
  maxSuppSize = maxSuppSize,
  nGamma = 5,
  gammaMin = 1e-04,
  gammaMax = 10
)
}
\arguments{
\item{X}{The rows are samples and the columns are genes of the matrix}

\item{Y}{The vector}

\item{penalty}{penalty = penalty}

\item{nFolds}{nFolds = 10}

\item{seed}{seed = 1}

\item{maxSuppSize}{maxSuppSize = maxSuppSize}

\item{nGamma}{nGamma = 5}

\item{gammaMin}{gammaMin = 0.0001}

\item{gammaMax}{gammaMax = 10}
}
\value{
A vector
}
\description{
inferCSN.fit
}
\examples{
inferCSN.fit(X, Y)
}
