% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.plot.R
\name{inferCSN.plot.dynamic.network}
\alias{inferCSN.plot.dynamic.network}
\title{Plot of dynamic networks}
\usage{
inferCSN.plot.dynamic.network(
  weightLists,
  tfs = NULL,
  onlyTFs = TRUE,
  order = NULL,
  thresh = NULL
)
}
\arguments{
\item{weightLists}{A list of ggplot2 objects}

\item{tfs}{tfs = NULL}

\item{onlyTFs}{onlyTFs = TRUE}

\item{order}{order = NULL}

\item{thresh}{thresh = NULL}
}
\value{
A list of ggplot2 objects
}
\description{
Plot of dynamic networks
}
