% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.R
\name{inferCSN}
\alias{inferCSN}
\title{inferCSN}
\usage{
inferCSN(
  data = NULL,
  normalize = FALSE,
  penalty = NULL,
  crossValidation = FALSE,
  nFolds = 10,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  nGamma = 5,
  verbose = FALSE,
  cores = 1
)
}
\arguments{
\item{data}{[Default = NULL] An object}

\item{normalize}{[Default = FALSE] Data normalize}

\item{penalty}{[Default = NULL]}

\item{crossValidation}{[Default = FALSE] Check whether cross validation is used}

\item{nFolds}{[Default = 10] N folds cross validation}

\item{regulators}{[Default = NULL] Regulator genes}

\item{targets}{[Default = NULL] Target genes}

\item{maxSuppSize}{[Default = NULL] The number of non-zore coef}

\item{nGamma}{[Default = 5] nGamma}

\item{verbose}{[Default = FALSE] Print detailed information}

\item{cores}{[Default = 1] CPU cores}
}
\value{
A list of gene-gene regulatory relationship
}
\description{
A method for inferring cell-type-specific gene regulatory network
from single-cell transcriptome data.
}
\examples{
data("exampleDataMatrix")
weightList <- inferCSN(exampleDataMatrix, verbose = TRUE)
}
