% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.R
\name{inferCSN}
\alias{inferCSN}
\title{inferCSN}
\usage{
inferCSN(
  matrix,
  penalty = NULL,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  cores = 1
)
}
\arguments{
\item{cores}{}
}
\description{
inferCSN
}
