% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.R
\name{inferCSN}
\alias{inferCSN}
\title{inferCSN
   A method for inferring cell-specific gene regulatory network from single-cell transcriptome data.}
\usage{
inferCSN(
  data = NULL,
  normalize = FALSE,
  penalty = NULL,
  crossValidation = FALSE,
  nFolds = 10,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  nGamma = 5,
  verbose = FALSE,
  cores = 1
)
}
\arguments{
\item{data}{An object}

\item{normalize}{Data normalize}

\item{penalty}{Default "L0"}

\item{crossValidation}{Cross validation}

\item{nFolds}{N folds cross validation}

\item{regulators}{Regulator genes}

\item{targets}{Target genes}

\item{maxSuppSize}{The number of non-zore coef}

\item{nGamma}{nGamma}

\item{verbose}{Print detailed information}

\item{cores}{CPU cores}
}
\value{
A list of gene-gene regulatory relationship
}
\description{
inferCSN
   A method for inferring cell-specific gene regulatory network from single-cell transcriptome data.
}
\examples{
  data("exampleDataMatrix")
  weightList <- inferCSN(exampleDataMatrix)
}
