% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.R
\name{inferCSN}
\alias{inferCSN}
\title{inferCSN}
\usage{
inferCSN(
  matrix,
  penalty = NULL,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  cores = 1
)
}
\arguments{
\item{matrix}{Gene experssion matrix}

\item{penalty}{Default "L0"}

\item{regulators}{Regulator genes}

\item{targets}{Target genes}

\item{maxSuppSize}{The number of non-zore coef}

\item{cores}{CPU cores}
}
\value{
A list of gene-gene regulatory relationship
}
\description{
inferCSN
}
\examples{
inferCSN(matrix)
}
