% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.R
\name{inferCSN}
\alias{inferCSN}
\alias{inferCSN,data.frame-method}
\alias{inferCSN,}
\alias{data.frame-method}
\alias{inferCSN,matrix-method}
\alias{matrix-method}
\title{Inferring cell-specific gene regulatory network}
\usage{
inferCSN(
  matrix,
  penalty = "L0",
  algorithm = "CD",
  crossValidation = FALSE,
  nFolds = 10,
  kFolds = NULL,
  rThreshold = 0,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  verbose = FALSE,
  cores = 1
)

\S4method{inferCSN}{data.frame}(
  matrix,
  penalty = "L0",
  algorithm = "CD",
  crossValidation = FALSE,
  nFolds = 10,
  kFolds = NULL,
  rThreshold = 0,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  verbose = FALSE,
  cores = 1
)

\S4method{inferCSN}{matrix}(
  matrix,
  penalty = "L0",
  algorithm = "CD",
  crossValidation = FALSE,
  nFolds = 10,
  kFolds = NULL,
  rThreshold = 0,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  verbose = FALSE,
  cores = 1
)
}
\arguments{
\item{matrix}{An expression matrix, cells by genes}

\item{penalty}{The type of regularization.
This can take either one of the following choices: "L0"and "L0L2".
For high-dimensional and sparse data, such as single-cell transcriptome data, "L0L2" is more effective}

\item{algorithm}{Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time}

\item{crossValidation}{Check whether cross validation is used}

\item{nFolds}{The number of folds for cross-validation}

\item{kFolds}{The number of folds for sample split}

\item{rThreshold}{rThreshold}

\item{regulators}{Regulator genes}

\item{targets}{Target genes}

\item{maxSuppSize}{The number of non-zore coef, this value will affect the final performance}

\item{verbose}{Print detailed information}

\item{cores}{CPU cores}
}
\value{
A data table of gene-gene regulatory relationship
}
\description{
Inferring cell-specific gene regulatory network
}
\examples{
library(inferCSN)
data("exampleMatrix")
weightDT <- inferCSN(exampleMatrix, verbose = TRUE)
head(weightDT)
weightDT <- inferCSN(exampleMatrix, cores = 2)
head(weightDT)

}
