% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.R
\name{inferCSN}
\alias{inferCSN}
\title{Inferring cell-specific gene regulatory network}
\usage{
inferCSN(
  matrix = NULL,
  penalty = NULL,
  algorithm = NULL,
  crossValidation = FALSE,
  nFolds = 10,
  regulators = NULL,
  targets = NULL,
  maxSuppSize = NULL,
  verbose = FALSE,
  cores = 1
)
}
\arguments{
\item{matrix}{An expression matrix, cells by genes}

\item{penalty}{[Default = "L0"] The type of regularization.
This can take either one of the following choices: "L0"and "L0L2".
For high-dimensional and sparse data, such as single-cell transcriptome data, "L0L2" is more effective}

\item{algorithm}{[Default = "CD"] Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time}

\item{crossValidation}{[Default = FALSE] Check whether cross validation is used}

\item{nFolds}{[Default = 10] The number of folds for cross-validation}

\item{regulators}{[Default = NULL] Regulator genes}

\item{targets}{[Default = NULL] Target genes}

\item{maxSuppSize}{[Default = NULL] The number of non-zore coef, this value will affect the final performance}

\item{verbose}{[Default = FALSE] Print detailed information}

\item{cores}{[Default = 1] CPU cores}
}
\value{
A data table of gene-gene regulatory relationship
}
\description{
Inferring cell-specific gene regulatory network
}
\examples{
library(inferCSN)
data("exampleMatrix")
weightDT <- inferCSN(exampleMatrix, verbose = TRUE)
head(weightDT)
weightDT <- inferCSN(exampleMatrix, cores = 2)
head(weightDT)

}
