% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN-package.R
\name{inferCSN_logo}
\alias{inferCSN_logo}
\title{inferCSN logo}
\usage{
inferCSN_logo(unicode = cli::is_utf8_output())
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE} on UTF-8 platforms.}
}
\description{
The inferCSN logo, using ASCII or Unicode characters
Use \code{\link[cli:ansi_strip]{cli::ansi_strip()}} to get rid of the colors.
}
\examples{
inferCSN_logo()
}
\references{
\url{https://github.com/tidyverse/tidyverse/blob/main/R/logo.R}
}
