% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cells.R
\name{meta_cells}
\alias{meta_cells}
\title{Detection of metacells from single-cell gene expression matrix}
\usage{
meta_cells(
  matrix,
  genes_use = NULL,
  genes_exclude = NULL,
  var_genes_num = min(1000, nrow(matrix)),
  gamma = 10,
  knn_k = 5,
  do_scale = TRUE,
  pc_num = 25,
  fast_pca = TRUE,
  do_approx = FALSE,
  approx_num = 20000,
  directed = FALSE,
  use_nn2 = TRUE,
  seed = 1,
  cluster_method = "walktrap",
  block_size = 10000,
  weights = NULL,
  do_median_norm = FALSE,
  ...
)
}
\arguments{
\item{matrix}{Log-normalized gene expression matrix with rows as genes and columns as cells.}

\item{genes_use}{A vector of genes used to compute PCA, default is \emph{\code{NULL}}.}

\item{genes_exclude}{A vector of genes to be excluded when computing PCA, default is \emph{\code{NULL}}.}

\item{var_genes_num}{If \emph{\code{genes_use}} is not provided, \emph{\code{var_genes_num}} genes with the largest
variation are used. Default is \emph{\code{min(1000, nrow(matrix))}}.}

\item{gamma}{Graining level of data (proportion of number of single cells in the initial dataset
to the number of metacells in the final dataset), default is \emph{\code{10}}.}

\item{knn_k}{Parameter to compute single-cell kNN network, default is \emph{\code{5}}.}

\item{do_scale}{Logical value, default is \emph{\code{TRUE}}, whether to scale gene expression matrix
when computing PCA.}

\item{pc_num}{Number of principal components to use for construction of single-cell kNN network,
default is \emph{\code{25}}.}

\item{fast_pca}{Logical value, default is \emph{\code{TRUE}}, use \link[irlba]{irlba} as a faster version
of prcomp.}

\item{do_approx}{Logical value, default is \emph{\code{FALSE}}, compute approximate kNN in case of a large
dataset (>50'000).}

\item{approx_num}{Number of cells to subsample for an approximate approach, default is \emph{\code{20000}}.}

\item{directed}{Logical value, default is \emph{\code{FALSE}}, whether to use directed graph.}

\item{use_nn2}{Logical value, default is \emph{\code{TRUE}}, whether to use nn2 function.}

\item{seed}{Seed to use to subsample cells for an approximate approach, default is \emph{\code{1}}.}

\item{cluster_method}{Clustering method to identify metacells, default is \emph{\code{walktrap}}.
Available methods are \emph{\code{walktrap}} and \emph{\code{louvain}} (not recommended, gamma is ignored).}

\item{block_size}{Number of cells to map to the nearest metacell at a time (for approx
coarse-graining), default is \emph{\code{10000}}.}

\item{weights}{Vector of cell weights (NULL by default),
used for computing average gene expression within cluster of metacells.}

\item{do_median_norm}{Logical value, default is \emph{\code{FALSE}},
whether to normalize by median value.}

\item{...}{Additional parameters passed to other methods.}
}
\value{
A metacell matrix where rows represent metacells and columns represent genes.
}
\description{
This function detects metacells from a single-cell gene expression matrix
using dimensionality reduction and clustering techniques.
}
\examples{
data("example_matrix")
meta_cells_matrix <- meta_cells(
  example_matrix,
  fast_pca = FALSE
)
dim(meta_cells_matrix)
meta_cells_matrix[1:6, 1:6]
}
\references{
https://github.com/GfellerLab/SuperCell
https://github.com/kuijjerlab/SCORPION
}
