% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN.plot.R
\name{net.format}
\alias{net.format}
\title{net.format}
\usage{
net.format(weightList, regulators = regulators)
}
\arguments{
\item{weightList}{weightList of GRN}

\item{regulators}{Regulators list}
}
\value{
A formated weight list
}
\description{
net.format
}
