% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{net.format}
\alias{net.format}
\title{Format weight table}
\usage{
net.format(weightDT, regulators = NULL)
}
\arguments{
\item{weightDT}{The weight data table of network}

\item{regulators}{Regulators list}
}
\value{
Format weight table
}
\description{
Format weight table
}
