% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalization}
\alias{normalization}
\title{Normalize numeric vector}
\usage{
normalization(x, method = "max_min")
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{Method for normalization.}
}
\value{
Normalized vector
}
\description{
Normalize numeric vector
}
