% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallelize_fun}
\alias{parallelize_fun}
\title{Parallelize a function}
\usage{
parallelize_fun(x, fun, cores = 1, export_fun = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A vector or list to apply over.}

\item{fun}{The function to be applied to each element.}

\item{cores}{The number of cores to use for parallelization with \code{\link[foreach]{foreach}}, default is \emph{\code{1}}.}

\item{export_fun}{The functions to export the function to workers.}

\item{verbose}{Logical value, default is \emph{\code{TRUE}}, whether to print progress messages.}
}
\value{
A list of computed results
}
\description{
Parallelize a function
}
