% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_contrast_networks}
\alias{plot_contrast_networks}
\title{Plot contrast networks}
\usage{
plot_contrast_networks(
  network_table,
  degree_value = 0,
  weight_value = 0,
  legend_position = "bottom"
)
}
\arguments{
\item{network_table}{The weight data table of network.}

\item{degree_value}{Degree value to filter nodes.}

\item{weight_value}{Weight value to filter edges.}

\item{legend_position}{The position of legend.}
}
\value{
A ggplot2 object
}
\description{
Plot contrast networks
}
\examples{
data("example_matrix")
network_table <- inferCSN(example_matrix)
plot_contrast_networks(network_table[1:50, ])
}
