% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_embedding}
\alias{plot_embedding}
\title{plot_embedding}
\usage{
plot_embedding(
  expression_matrix,
  labels = NULL,
  method = "tsne",
  colors = RColorBrewer::brewer.pal(length(unique(labels)), "Set1"),
  seed = 1,
  point_size = 1
)
}
\arguments{
\item{expression_matrix}{Input data frame.}

\item{labels}{Input data frame.}

\item{method}{Method to use for dimensionality reduction.}

\item{colors}{Colors to use for the plot.}

\item{seed}{Seed for the random number generator.}

\item{point_size}{Size of the points.}
}
\value{
ggplot object
}
\description{
Plot embedding of the expression matrix.
}
\examples{
data("example_matrix")
samples_use <- 1:200
labels <- sample(
  c("A", "B", "C", "D", "E"),
  nrow(example_matrix[samples_use, ]),
  replace = TRUE
)

plot_embedding(
  example_matrix[samples_use, ],
  point_size = 2
)

plot_embedding(
  example_matrix[samples_use, ],
  labels,
  point_size = 2
)

plot_embedding(
  example_matrix[samples_use, ],
  labels,
  method = "pca",
  point_size = 2
)
}
