% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.inferCSN}
\alias{print.inferCSN}
\alias{print.inferCSNCV}
\title{Prints a summary of inferCSN.fit}
\usage{
\method{print}{inferCSN}(x, ...)

\method{print}{inferCSNCV}(x, ...)
}
\arguments{
\item{x}{The output of inferCSN.fit or inferCSN.cvfit}

\item{...}{Other parameters}
}
\value{
Return the information of inferCSN.fit

Return the information of inferCSN.fit
}
\description{
Prints a summary of inferCSN.fit
}
