% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.srm}
\alias{print.srm}
\alias{print.srm_cv}
\title{Prints a summary of \code{fit_sparse_regression}}
\usage{
\method{print}{srm}(x, ...)

\method{print}{srm_cv}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link{fit_sparse_regression}}.}

\item{...}{Other parameters}
}
\value{
Return information of \code{fit_sparse_regression}
}
\description{
Prints a summary of \code{fit_sparse_regression}
}
