% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.SRM_fit}
\alias{print.SRM_fit}
\alias{print.SRM_fit_CV}
\title{Prints a summary of model.fit}
\usage{
\method{print}{SRM_fit}(x, ...)

\method{print}{SRM_fit_CV}(x, ...)
}
\arguments{
\item{x}{The output of model.fit or inferCSN.cvfit}

\item{...}{Other parameters}
}
\value{
Return information of model.fit

Return information of model.fit
}
\description{
Prints a summary of model.fit
}
