% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.regression.R
\name{sparse.regression}
\alias{sparse.regression}
\title{sparse.regression}
\usage{
sparse.regression(
  X,
  y,
  crossValidation = FALSE,
  penalty = "L0",
  algorithm = "CD",
  maxSuppSize = NULL,
  nFolds = 10,
  nGamma = 5,
  verbose = verbose
)
}
\arguments{
\item{X}{The data matrix}

\item{y}{The response vector}

\item{crossValidation}{[Default = FALSE] Check whether cross validation is used}

\item{penalty}{[Default = "L0"] The type of regularization
This can take either one of the following choices: "L0"and "L0L2"
For high-dimensional and sparse data, such as single-cell transcriptome data, "L0L2" is more effective}

\item{algorithm}{[Default = "CD"] Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time}

\item{maxSuppSize}{[Default = NULL] The number of non-zore coef, this value will affect the final performance}

\item{nFolds}{[Default = 10] N folds cross validation}

\item{nGamma}{[Default = 5] nGamma}

\item{verbose}{[Default = FALSE] Print detailed information}
}
\value{
A vector of weight
}
\description{
sparse.regression
}
