% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.regression.R
\name{sparse.regression}
\alias{sparse.regression}
\title{Sparse regression model}
\usage{
sparse.regression(
  X,
  y,
  crossValidation = FALSE,
  seed = 1,
  penalty = "L0",
  algorithm = "CD",
  maxSuppSize = NULL,
  nFolds = 10,
  kFolds = NULL,
  rThreshold = 0,
  verbose = FALSE
)
}
\arguments{
\item{X}{The data matrix}

\item{y}{The response vector}

\item{crossValidation}{Check whether cross validation is used.}

\item{seed}{The seed used in randomly shuffling the data for cross-validation.}

\item{penalty}{The type of regularization.
This can take either one of the following choices: "L0" and "L0L2".
For high-dimensional and sparse data, such as single-cell sequencing data, "L0L2" is more effective.}

\item{algorithm}{The type of algorithm used to minimize the objective function.
Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time.}

\item{maxSuppSize}{The number of non-zore coef, this value will affect the final performance.
The maximum support size at which to terminate the regularization path.
Recommend setting this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small portion of non-zeros.}

\item{nFolds}{The number of folds for cross-validation.}

\item{kFolds}{The number of folds for sample split.}

\item{rThreshold}{Threshold of R^2.}

\item{verbose}{Print detailed information.}
}
\value{
The coefficients
}
\description{
Sparse regression model
}
