% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_regression.R
\name{sparse_regression}
\alias{sparse_regression}
\title{Sparse regression model}
\usage{
sparse_regression(
  x,
  y,
  cross_validation = FALSE,
  seed = 1,
  penalty = "L0",
  algorithm = "CD",
  regulators_num = ncol(x),
  n_folds = 10,
  percent_samples = 1,
  r_threshold = 0,
  computation_method = "cor",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{The data matrix}

\item{y}{The response vector}

\item{cross_validation}{Check whether cross validation is used.}

\item{seed}{The seed used in randomly shuffling the data for cross-validation.}

\item{penalty}{The type of regularization.
This can take either one of the following choices: \code{L0}, \code{L0L1} and \code{L0L2}.
For high-dimensional and sparse data, such as single-cell sequencing data, \code{L0L2} is more effective.}

\item{algorithm}{The type of algorithm used to minimize the objective function.
Currently \code{CD} and \code{CDPSI} are supported.
The \code{CDPSI} algorithm may yield better results, but it also increases running time.}

\item{regulators_num}{The number of non-zore coefficients, this value will affect the final performance.
The maximum support size at which to terminate the regularization path.
Recommend setting this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small portion of non-zeros.}

\item{n_folds}{The number of folds for cross-validation.}

\item{percent_samples}{The percent of all samples used for \code{\link{sparse_regression}}. Default set to 1.}

\item{r_threshold}{Threshold of \eqn{R^2} or correlation coefficient.}

\item{computation_method}{The method used to compute \code{r}.}

\item{verbose}{Logical value. Whether to print detailed information.}

\item{...}{Parameters for other methods.}
}
\value{
Coefficients
}
\description{
Sparse regression model
}
\examples{
data("example_matrix")
sparse_regression(
  example_matrix[, -1],
  example_matrix[, 1]
)
}
