% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table.to.matrix}
\alias{table.to.matrix}
\title{Switch weight table to matrix}
\usage{
table.to.matrix(weight_table)
}
\arguments{
\item{weight_table}{The weight data table of network.}
}
\value{
Weight matrix
}
\description{
Switch weight table to matrix
}
