% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_to_matrix}
\alias{table_to_matrix}
\title{Switch network table to matrix}
\usage{
table_to_matrix(network_table, regulators = NULL, targets = NULL)
}
\arguments{
\item{network_table}{The weight data table of network.}

\item{regulators}{Regulators list.}

\item{targets}{Targets list.}
}
\value{
Weight matrix
}
\description{
Switch network table to matrix
}
\examples{
data("example_matrix")
network_table <- inferCSN(example_matrix)
head(network_table)

table_to_matrix(network_table)[1:6, 1:6]

table_to_matrix(
  network_table,
  regulators = c("g1", "g2"),
  targets = c("g3", "g4")
)
}
